Add ThemisDB Wiki Integration plugin and documentation importer
- Introduced ThemisDB Wiki Integration plugin for fetching and displaying documentation from GitHub in WordPress.
- Implemented shortcodes for wiki content, documentation listing, and navigation.
- Added admin settings for configuration and AJAX handlers for cache management.
- Created a WordPress documentation importer that utilizes a JSON file for intelligent category and tag extraction.
- Included WP-CLI support for running the importer and added an admin interface for manual imports.

+++ include/query/workload_cache_strategy.h
@@ -55,7 +55,7 @@ struct QueryCharacteristics {
             last_accessed - first_seen);
         // Handle case where query was just added (duration ~0)
         // Use minimum duration of 1 minute to avoid division by zero
-        int64_t duration_min = std::max(duration.count(), int64_t(1));
+        int64_t duration_min = std::max(static_cast<int64_t>(duration.count()), int64_t(1));
         return static_cast<double>(access_count) / duration_min;
     }
 };
